test_shell_v4

+++ smarthouse_v4/src/smarthouse_host/smarthouse_shell.c
@@ -33,10 +33,10 @@ int ledOffFn(void)
 
 int ledOnFn(void)
 {
-	uint8_t set_pin=setPin();
+	int set_pin=setPin();
 	if(set_pin<=-1)
 	{	return -1;	}
-	digital_config.pin_digital=set_pin;
+	digital_config.pin_digital=(uint8_t)set_pin;
 	digital_config.set_digital=ledOn;
 	pointer_packet=(PacketHeader*)&digital_config;
 	return 0;
